Add UUID for game instances
Without proper lobby, games are joined and created randomly. For this
purpose we do not need unique identifiers for game instances. However,
as we are going to implement an lobby system, we will need to have some
kind of identifier for games that can be joined and UUID is just great
for that.