Add factory for different game user interfaces
New function in `game-factory.js` is now capable of constructing user
interface for a game based only on the game mode and websocket.
Currently it has only Panel League duel because it's the only game mode
which is "complete" and other game modes are somewhat work in progress
in other branches.